requirements.txt
streamlit
google-generativeai
langchain
chromadb
python-dotenv
PyPDF2
scikit-learn
dotenv
langchain-community
fpdf
langchain_google_genai
faiss-cpu
rank_bm25
fastapi
uvicorn
jinja2
python-docx
docx2pdf
python-multipart
pydantic
requests